Continued developing UX and refactored react code base
