+++ client/src/DownloadPage/DownloadPage.css
@@ -1,0 +1,33 @@
+/* DownloadPage.css */
+.return-button {
+  margin-top: 124px;
+  border: none;
+}
+
+.download-container {
+  position: relative;
+  margin-top: 0px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  height: 10vh; /* 100% of the viewport height */
+}
+
+.download-link {
+    display: inline-block;
+    padding: 10px 20px;
+    background-color: rgb(255, 255, 255, 0.7); /* Background color */
+    color: black; /* Text color */
+    text-shadow: 2px 2px 4px rgb(0, 0, 0, 0.2);
+    border: 5px black;
+    text-decoration: none;
+    border-radius: 5px;
+    font-weight: bold;
+    transition: background-color 0.3s ease; /* Smooth background color transition */
+}
+  
+.download-link:hover {
+  background-color: white; /* Change the background color on hover */
+  text-decoration: underline; /* Underline the text on hover */
+  cursor: pointer; /* Show a pointer cursor on hover */
+}
