+++ client/src/DownloadPage/Spinner.css
@@ -1,0 +1,14 @@
+/* Spinner.css */
+.spinner {
+    border: 4px solid rgba(255, 255, 255, 0.3);
+    border-radius: 50%;
+    border-top: 4px solid #007bff; /* Adjust the color as needed */
+    width: 40px;
+    height: 40px;
+    animation: spin 1s linear infinite; /* CSS animation */
+  }
+  
+  @keyframes spin {
+    0% { transform: rotate(0deg); }
+    100% { transform: rotate(360deg); }
+  }
