+++ client/src/DownloadPage/Spinner.js
@@ -1,0 +1,9 @@
+// Spinner.js
+import React from 'react';
+import './Spinner.css';
+
+const Spinner = () => {
+  return <div className="spinner"></div>;
+};
+
+export default Spinner;
