+++ client/src/Header/Header.js
@@ -4,9 +4,9 @@ import GPTComponent from './GPTComponent';
 const Header = () => {
     return (
         <header className="site-header">
-            <h1>AI Powered Inventory Tracker</h1>
+            <h1>Inventory Spreadsheet Generator</h1>
             <div className="powered-by-container"> {/* Flex container */}
-                <p>You give commands. We return you a csv of everything you told. Powered by chatGPT </p>
+                <p>AI infused Inventory tool. Powered by ChatGPT </p>
                 <GPTComponent /> {/* Assuming this renders the symbol */}
             </div>
         </header>
