+++ client/src/MainPage/App.css
@@ -41,10 +41,6 @@
   margin-bottom: 0px;
 }
 
-.header-text {
-  text-shadow: 1px 1px 8px rgba(0, 0, 0, 0.2);
-}
-
 /* Add these styles to center the container */
 .title-container {
   position: relative;
@@ -109,16 +105,11 @@ textarea:focus {
 .my-button {
     position: relative; /* or absolute, fixed, depending on use case */
     bottom: 100px;
-    background: linear-gradient(
-        to bottom, 
-        rgba(255, 255, 255, .7) 15%, /* Start with full opacity */
-        rgba(255, 255, 255, .45) 70% /* End with full transparency */
-    );
-    color: black; /* White text */
+    background: rgb(255, 255, 255, 0.7);
+    color: black;
     padding: 15px 32px; /* sPadding around the text */
-    border: 1.9px solid #000E23; 
     color: black; /* Lighter text color for the description */
-    font-size: 1.2em;
+    font-size: 1.1em;
     justify-content: center; /* Center horizontally */
     border-radius: 12px; /* Rounded corners */
     cursor: pointer; /* Pointer cursor on hover */
@@ -129,7 +120,7 @@ textarea:focus {
   }
   
 .my-button:hover {
-  background-color: rgba(255, 255, 255, 0); 
+  background-color: rgba(255, 255, 255, 1); 
   border: 2px solid black;
   text-shadow: 1px 1px 8px rgba(0, 0, 0, 0.3);
   box-shadow: 8px 8px 12px rgba(0, 0, 0, 0.3); /* Glowing effect on focus */
