+++ client/src/styling.css
@@ -17,3 +17,7 @@ body {
   opacity: 1;
   z-index: -1;
 }
+
+.header-text {
+  text-shadow: 2px 2px 8px rgba(0, 0, 0, 0.2);
+}
